Hide noop proxy monitor
